requirements.txt
pypdf==4.3.1
langchain==0.2.16
langchain-google-genai==1.0.10
gradio==4.44.1
python-dotenv==1.0.1
faiss-cpu==1.8.0
langchain-community
streamlit==1.38.0
requests==2.32.3
serpapi==0.1.5